added own vector(way better than glm performance and api vise) added types and refactored to use new types and maths

+++ Jin/OrthographicCamera.h
@@ -4,7 +4,7 @@
 class OrthographicCamera : public Camera
 {	
 public:
-	OrthographicCamera(int width, int height);
+	OrthographicCamera(f32 width, f32 height);
 
 	
 	virtual void Tick() override;
